Make divmod's signature generic in preparation for u8 and u64.
commit-id:06dbc033
