+++ crates/cairo-lang-sierra-ap-change/src/core_libfunc_ap_change.rs
@@ -136,7 +136,7 @@ pub fn core_libfunc_ap_change<InfoProvider: InvocationApChangeInfoProvider>(
                     vec![ApChange::Known(2), ApChange::Known(3)]
                 }
             },
-            Uint128Concrete::DivMod(_) => vec![ApChange::Known(7)],
+            Uint128Concrete::Divmod(_) => vec![ApChange::Known(7)],
             Uint128Concrete::WideMul(_) => vec![ApChange::Known(17)],
             Uint128Concrete::LessThan(_) => vec![ApChange::Known(2), ApChange::Known(3)],
             Uint128Concrete::Equal(_) => vec![ApChange::Known(1), ApChange::Known(1)],
